frame pack & unpack

+++ source/powerwriter_at_core/powerwriter_at_porting.c
@@ -59,7 +59,7 @@ extern "C"
 #endif
 
 /* crc */
-uint16_t	GetCrc32(void * data, size_t size)
+uint32_t	GetCrc32(void * data, size_t size)
 {
 #ifdef POWERWRITER_AT_HWCRC_ENABLE
 	/*
