+++ source/powerwriter_at_core/powerwriter_at_porting.h
@@ -30,7 +30,7 @@ extern "C" {
 #endif
 
 uint32_t	GetSystemTick(void);									/* Get System Tick */
-uint16_t	GetCrc32(void * data, size_t size);		/* Get crc16 */
+uint32_t	GetCrc32(void * data, size_t size);		/* Get crc16 */
 
 #ifdef __cplusplus
 }
